// app/src/api/model/missingValueHandlingType.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Economicon App API
 * OpenAPI spec version: 0.2.0
 */

export type MissingValueHandlingType = typeof MissingValueHandlingType[keyof typeof MissingValueHandlingType];


export const MissingValueHandlingType = {
  ignore: 'ignore',
  remove: 'remove',
  error: 'error',
} as const;