// app/src/api/model/robustStandardErrorHcType.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Economicon App API
 * OpenAPI spec version: 0.1.0
 */

export type RobustStandardErrorHcType = typeof RobustStandardErrorHcType[keyof typeof RobustStandardErrorHcType];


export const RobustStandardErrorHcType = {
  HC0: 'HC0',
  HC1: 'HC1',
  HC2: 'HC2',
  HC3: 'HC3',
} as const;